requirements.txt
absl-py==1.4.0
addict==2.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
# certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
chumpy==0.70
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
coverage==6.5.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.4
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.0.4
fastjsonschema==2.16.2
flake8==5.0.4
flatbuffers==22.12.6
# flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
freetype-py==2.3.0
# fvcore @ git+https://github.com/facebookresearch/fvcore@d39b53de2c1e363e2ceb27d321d403b4bdb2b7a7
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
humanfriendly==10.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==6.0.0
importlib-resources==5.10.1
iniconfig==1.1.1
interrogate==1.5.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isort==4.3.21
jedi==0.18.2
Jinja2==3.1.2
json-tricks==3.16.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mkl-fft==1.3.1
# mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mmcv-full==1.5.0
-e git+https://github.com/open-mmlab/mmdetection.git@31c84958f54287a8be2b99cbf87a6dcf12e57753#egg=mmdet
-e git+https://github.com/open-mmlab/mmpose.git@148091e7a021c0f2f2e7060f945f76640e06263d#egg=mmpose
motmetrics==1.2.5
moviepy==1.0.3
mpmath==1.2.1
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
oauthlib==3.2.2
onnx==1.13.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==22.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
portalocker==2.6.0
poseval @ git+https://github.com/svenkreiss/poseval.git@3128c5cbcf90946e5164ff438ad651e113e64613
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pymemcache==4.0.0
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
pyrsistent==0.19.2
-e git+https://github.com/kennymckormick/pyskl.git@bd46e966ab0e332a5cd3a806ae319e6fd4df9ed0#egg=pyskl
pytest==7.2.0
pytest-runner==6.0.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smplx==0.1.28
sniffio==1.3.0
soupsieve==2.3.2.post1
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
terminado==0.17.1
terminaltables==3.1.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trimesh==3.17.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
widgetsnbextension==4.0.4
xdoctest==1.1.0
xmltodict==0.13.0
xtcocotools==1.12
yacs==0.1.8
yapf==0.32.0
zipp==3.11.0
